refactor: give project-ignore its own directory, and say something a person can do
Twenty-seven rounds of correction left one bounded component split between its
named owner and hook-utils, which is the content-blank name the rules single out
precisely because it absorbs whatever fits — and it was the half that kept growing.
The component now lives in clients/hooks/lib/project-ignore/, with the arena and
lock, the durability marker lifecycle, the legacy sweep and the shared artifact
primitives each in their own module; the core drops from 1790 lines to 988 and
hook-utils from 515 to 357. Nothing moved that another hook uses, and nothing
changed but import paths and names the directory already scopes.

Three outcomes an operator can actually reach — killed, no output, unreadable
output — described what had happened and stopped there, and the arena remedies said
to repair <commonGitDir>/coral, which is a placeholder rather than a path. They now
name a command to run and a component to fix. docs/hooks.md had four unrelated
behaviours compressed into five dense bullets with no destination for the words the
remedies use; it has a section that introduces the arena, the migration, the
preconditions, the refusals, and what quarantined evidence is and where it sits.

Tests that policed prose for banned words, counted occurrences, and pinned the
shell's branch text went; the exit-code literals stayed pinned to their constants,
because the snippet cannot import them. The validator no longer launches a
subprocess on all 137 helper calls. Two refusal paths that only hand-built fixtures
had ever reached now have real filesystem coverage. And the file that had long since
stopped being about symlinks is named for what it tests.
