+++ clients/hooks/lib/hook-utils.mjs
@@ -1,19 +1,5 @@
 import { execSync } from 'node:child_process';
-import {
-  chmodSync,
-  closeSync,
-  constants,
-  fchmodSync,
-  fstatSync,
-  lstatSync,
-  mkdirSync,
-  openSync,
-  readdirSync,
-  readFileSync,
-  realpathSync,
-  statSync,
-  unlinkSync,
-} from 'node:fs';
+import { readdirSync, readFileSync, statSync, unlinkSync } from 'node:fs';
 import { homedir } from 'node:os';
 import { basename, dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
@@ -332,150 +318,6 @@ export function claudeConfigDir() {
 // request and never select a different daemon or state tree.
 export function coralStateRoot() {
   return join(homedir(), '.coral');
-}
-
-export const PROJECT_IGNORE_SPAWN_TIMEOUT_MS = 5000;
-export const PROJECT_IGNORE_CONTEXT_PROBE_BUDGET_MS = 1500;
-export const PROJECT_IGNORE_LOCK_WRAPPER_BUDGET_MS = 250;
-export const PROJECT_IGNORE_STAGING_ARENA_MAX_AGE_MS = 600_000;
-export const PROJECT_IGNORE_ARENA_SWEEP_BUDGET_MS = 250;
-export const PROJECT_IGNORE_ARENA_SWEEP_MAX_RUNS = 32;
-export const PROJECT_IGNORE_LOCK_UNAVAILABLE_EXIT_CODE = 69;
-export const PROJECT_IGNORE_LOCK_CONFLICT_EXIT_CODE = 75;
-
-export function projectIgnoreContextProbeDeadline(startedNs) {
-  try {
-    return BigInt(startedNs) + BigInt(PROJECT_IGNORE_CONTEXT_PROBE_BUDGET_MS) * 1_000_000n;
-  } catch {
-    return null;
-  }
-}
-
-export function projectIgnoreStagingDir() {
-  return join(coralStateRoot(), 'staging', 'project-ignore');
-}
-
-export function projectIgnoreMaintenanceLockPath() {
-  return join(coralStateRoot(), 'staging', 'project-ignore.maintenance.lock');
-}
-
-function ensureRealDirectoryComponent(path) {
-  try {
-    const stat = lstatSync(path);
-    if (stat.isSymbolicLink() || !stat.isDirectory()) return { kind: 'structural-conflict' };
-    chmodSync(path, 0o700);
-    return { kind: 'prepared' };
-  } catch (error) {
-    if (error?.code !== 'ENOENT') return { kind: 'operational-failure' };
-  }
-
-  try {
-    mkdirSync(path, { mode: 0o700 });
-  } catch (error) {
-    if (error?.code !== 'EEXIST') return { kind: 'operational-failure' };
-  }
-
-  try {
-    const stat = lstatSync(path);
-    if (stat.isSymbolicLink() || !stat.isDirectory()) return { kind: 'structural-conflict' };
-    chmodSync(path, 0o700);
-    return { kind: 'prepared' };
-  } catch {
-    return { kind: 'operational-failure' };
-  }
-}
-
-function prepareCoralStateDirectory(components) {
-  try {
-    const stateRoot = coralStateRoot();
-    const home = realpathSync(dirname(stateRoot));
-    const expectedStateRoot = join(home, basename(stateRoot));
-    const stateRootPreparation = ensureRealDirectoryComponent(expectedStateRoot);
-    if (stateRootPreparation.kind !== 'prepared') return stateRootPreparation;
-    let canonicalStateRoot = realpathSync(stateRoot);
-    if (canonicalStateRoot !== expectedStateRoot) return { kind: 'structural-conflict' };
-    for (const component of components) {
-      const expectedComponent = join(canonicalStateRoot, component);
-      const componentPreparation = ensureRealDirectoryComponent(expectedComponent);
-      if (componentPreparation.kind !== 'prepared') return componentPreparation;
-      canonicalStateRoot = realpathSync(expectedComponent);
-      if (canonicalStateRoot !== expectedComponent) return { kind: 'structural-conflict' };
-    }
-    return { kind: 'prepared', path: canonicalStateRoot };
-  } catch {
-    return { kind: 'operational-failure' };
-  }
-}
-
-function prepareProjectIgnoreStateStagingDir() {
-  const preparation = prepareCoralStateDirectory(['staging']);
-  return preparation.kind === 'prepared' ? preparation.path : null;
-}
-
-export function prepareCoralProjectDir(target) {
-  const projectsRootName = buildFlavor() === 'dev' ? 'projects-dev' : 'projects';
-  const projectsRoot = join(coralStateRoot(), projectsRootName);
-  if (target === projectsRoot || dirname(target) !== projectsRoot) return { kind: 'structural-conflict' };
-  const projectName = basename(target);
-  if (join(projectsRoot, projectName) !== target) return { kind: 'structural-conflict' };
-  return prepareCoralStateDirectory([projectsRootName, projectName]);
-}
-
-export function prepareProjectIgnoreStagingDir() {
-  const stagingDir = prepareProjectIgnoreStateStagingDir();
-  if (!stagingDir) return null;
-  const arena = join(stagingDir, basename(projectIgnoreStagingDir()));
-  const arenaPreparation = ensureRealDirectoryComponent(arena);
-  if (arenaPreparation.kind !== 'prepared') return null;
-  try {
-    const canonicalArena = realpathSync(arena);
-    return canonicalArena === arena ? canonicalArena : null;
-  } catch {
-    return null;
-  }
-}
-
-export function openProjectIgnoreMaintenanceLock() {
-  const stagingDir = prepareProjectIgnoreStateStagingDir();
-  if (!stagingDir) return null;
-
-  const lockPath = projectIgnoreMaintenanceLockPath();
-  let fd;
-  try {
-    try {
-      const named = lstatSync(lockPath);
-      if (named.isSymbolicLink() || !named.isFile()) return null;
-      chmodSync(lockPath, 0o600);
-    } catch (error) {
-      if (error?.code !== 'ENOENT') return null;
-    }
-    fd = openSync(
-      lockPath,
-      constants.O_RDWR | constants.O_CREAT | (constants.O_NOFOLLOW ?? 0),
-      0o600,
-    );
-    const opened = fstatSync(fd);
-    const named = lstatSync(lockPath);
-    if (
-      !opened.isFile() ||
-      named.isSymbolicLink() ||
-      !named.isFile() ||
-      opened.dev !== named.dev ||
-      opened.ino !== named.ino
-    ) {
-      closeSync(fd);
-      return null;
-    }
-    fchmodSync(fd, 0o600);
-    return fd;
-  } catch {
-    if (fd !== undefined) {
-      try {
-        closeSync(fd);
-      } catch {}
-    }
-    return null;
-  }
 }
 
 export function resolveKbRoot() {
