+++ clients/hooks/lib/project-ignore/result.mjs
